fix bug: remove follower

+++ backend/socialdistribution/viewsets/friend_view.py
@@ -96,11 +96,16 @@ class FriendViewSet(viewsets.ModelViewSet):
     def remove_follower(self, request, author_id, foreign_author_id):
 
         id = f'{foreign_author_id}to{author_id}'
+        print(">>>>>>>>>>")
+        print(id)
          
         try:
             follower = FollowRequest.objects.get(id=id)
-            follower.delete()
-            response_msg = "Successfully delete"
+            if follower.relation == 'F':
+                follower.delete()
+                response_msg = "Successfully delete"
+            else:
+                response_msg = 'Not your follower'
         except:
             response_msg = 'Not your follower'
